third_party/blink/web_tests/external/wpt/payment-request/payment-request-hasenrolledinstrument-method-protection.tentative.https-expected.txt

This is a testharness.js-based test.
FAIL Optionally, at the user agent's discretion, return a promise rejected with a "NotAllowedError" DOMException. assert_equals: If it throws, then it must be a NotAllowedError. expected "NotAllowedError" but got "UnknownError"
Harness: the test ran to completion.
